+++ client/src/services/ProcuraHub.js
@@ -1,0 +1,59 @@
+// src/services/ProcuraHub.js
+const BASE_URL = "http://localhost:6000/api"; // your backend
+
+export async function register(registerData) {
+  const response = await fetch(`${BASE_URL}/auth/register/`, {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify({
+      fullName: registerData.fullName,
+      company: registerData.company,
+      email: registerData.workEmail,
+      password: registerData.password,
+      role: registerData.role,
+    }),
+  });
+
+  if (!response.ok) {
+    const errorText = await response.text();
+    throw new Error(errorText || `Request failed with status ${response.status}`);
+  }
+
+  return response.json();
+}
+
+export async function login(loginData) {
+  const response = await fetch(`${BASE_URL}/auth/login/`, {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify({
+      email: loginData.email,
+      password: loginData.password,
+    }),
+  });
+
+  if (!response.ok) {
+    const errorText = await response.text();
+    throw new Error(errorText || `Request failed with status ${response.status}`);
+  }
+
+  return response.json();
+}
+
+export async function createRequest(token, formData) {
+  const response = await fetch(`${BASE_URL}/requests/createRequest`, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/x-www-form-urlencoded",
+      Authorization: `Bearer ${token}`,
+    },
+    body: new URLSearchParams(formData),
+  });
+
+  if (!response.ok) {
+    const errorText = await response.text();
+    throw new Error(errorText || `Request failed with status ${response.status}`);
+  }
+
+  return response.json();
+}
